Apply philnik777's suggestions
