+++ libcxx/test/std/algorithms/alg.modifying.operations/alg.move/move_backward.pass.cpp
@@ -21,48 +21,24 @@
 #include <memory>
 
 #include "MoveOnly.h"
-#include "../alg.copy/common.h"
 #include "test_iterators.h"
 #include "test_macros.h"
 #include "type_algorithms.h"
 
-TEST_CONSTEXPR_CXX20 void test_padding() { // Make sure that padding bits aren't copied
-  Derived src(1, 2, 3);
-  Derived dst(4, 5, 6);
-  std::move_backward(
-      static_cast<PaddedBase*>(&src), static_cast<PaddedBase*>(&src) + 1, static_cast<PaddedBase*>(&dst) + 1);
-  assert(dst.a_ == 1);
-  assert(dst.b_ == 2);
-  assert(dst.c_ == 6);
-}
+class PaddedBase {
+public:
+  TEST_CONSTEXPR PaddedBase(std::int16_t a, std::int8_t b) : a_(a), b_(b) {}
 
-TEST_CONSTEXPR_CXX20 void test_overlapping() { // Make sure that overlapping ranges can be copied
-  int a[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-  std::move_backward(a, a + 7, a + 10);
-  int expected[] = {1, 2, 3, 1, 2, 3, 4, 5, 6, 7};
-  assert(std::equal(a, a + 10, expected));
-}
+  std::int16_t a_;
+  std::int8_t b_;
+};
 
-TEST_CONSTEXPR_CXX20 void test_moveonly() { // Make sure that the algorithm works with move-only types
-  // When non-trivial
-  {
-    MoveOnly from[3] = {1, 2, 3};
-    MoveOnly to[3]   = {};
-    std::move_backward(std::begin(from), std::end(from), std::end(to));
-    assert(to[0] == MoveOnly(1));
-    assert(to[1] == MoveOnly(2));
-    assert(to[2] == MoveOnly(3));
-  }
-  // When trivial
-  {
-    TrivialMoveOnly from[3] = {1, 2, 3};
-    TrivialMoveOnly to[3]   = {};
-    std::move_backward(std::begin(from), std::end(from), std::end(to));
-    assert(to[0] == TrivialMoveOnly(1));
-    assert(to[1] == TrivialMoveOnly(2));
-    assert(to[2] == TrivialMoveOnly(3));
-  }
-}
+class Derived : public PaddedBase {
+public:
+  TEST_CONSTEXPR Derived(std::int16_t a, std::int8_t b, std::int8_t c) : PaddedBase(a, b), c_(c) {}
+
+  std::int8_t c_;
+};
 
 template <class InIter>
 struct Test {
@@ -117,9 +93,41 @@ TEST_CONSTEXPR_CXX20 bool test() {
   types::for_each(types::bidirectional_iterator_list<int*>(), TestOutIters());
   if (TEST_STD_AT_LEAST_23_OR_RUNTIME_EVALUATED)
     types::for_each(types::bidirectional_iterator_list<std::unique_ptr<int>*>(), Test1OutIters());
-  test_padding();
-  test_overlapping();
-  test_moveonly();
+  { // Make sure that padding bits aren't copied
+    Derived src(1, 2, 3);
+    Derived dst(4, 5, 6);
+    std::move_backward(
+        static_cast<PaddedBase*>(&src), static_cast<PaddedBase*>(&src) + 1, static_cast<PaddedBase*>(&dst) + 1);
+    assert(dst.a_ == 1);
+    assert(dst.b_ == 2);
+    assert(dst.c_ == 6);
+  }
+  { // Make sure that overlapping ranges can be copied
+    int a[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+    std::move_backward(a, a + 7, a + 10);
+    int expected[] = {1, 2, 3, 1, 2, 3, 4, 5, 6, 7};
+    assert(std::equal(a, a + 10, expected));
+  }
+  { // Make sure that the algorithm works with move-only types
+    // When non-trivial
+    {
+      MoveOnly from[3] = {1, 2, 3};
+      MoveOnly to[3]   = {};
+      std::move_backward(std::begin(from), std::end(from), std::end(to));
+      assert(to[0] == MoveOnly(1));
+      assert(to[1] == MoveOnly(2));
+      assert(to[2] == MoveOnly(3));
+    }
+    // When trivial
+    {
+      TrivialMoveOnly from[3] = {1, 2, 3};
+      TrivialMoveOnly to[3]   = {};
+      std::move_backward(std::begin(from), std::end(from), std::end(to));
+      assert(to[0] == TrivialMoveOnly(1));
+      assert(to[1] == TrivialMoveOnly(2));
+      assert(to[2] == TrivialMoveOnly(3));
+    }
+  }
 
   return true;
 }
